// src/app/models/WDHours.ts
export class WDHours {
    detail_WD_HOURS_ID: number;
    date_WD: String;
    description: string;

    hour_1: number;
    hour_2: number;
    hour_3: number;
    hour_4: number;
    hour_5: number;
    hour_6: number;
    hour_7: number;
    hour_8: number;
    hour_9: number;
    hour_10: number;
    hour_11: number;
    hour_12: number;
    hour_13: number;
    hour_14: number;
    hour_15: number;
    hour_16: number;
    hour_17: number;
    hour_18: number;
    hour_19: number;
    hour_20: number;
    hour_21: number;
    hour_22: number;
    hour_23: number;
    hour_24: number;
    
    status: number;
    creation_DATE: Date;
    created_BY: string;
    last_UPDATE_DATE: Date;
    last_UPDATED_BY: string;
}