// src/app/models/WDHoursSpecific.ts
export class WDHoursSpecific {
    detail_WD_HOURS_SPECIFIC_ID: number
    date_WD: string
    description: string
    
    shift1_START_TIME: string
    shift1_END_TIME: string
    shift1_TOTAL_TIME: number
    
    shift2_START_TIME: string
    shift2_END_TIME: string
    shift2_TOTAL_TIME: number
    
    shift3_START_TIME: string
    shift3_END_TIME: string
    shift3_TOTAL_TIME: number
    
    status: number
    created_BY: string
    creation_DATE: Date
    last_UPDATED_BY: string
    last_UPDATE_DATE: Date
  }
  